Split SearchResultsView.vue into a component and a composable
